Add user config for custom GLPK location
This commit add user configuration to "setup.py" that allows for custom
GLPK install location. The user can either specify `--with-glpk=<path>`
or `GLPK=<path>` environment variable. If both are specified, the
command line option takes precedence. In both cases, it is assumed that
glpk is *not* located on the C compiler's library path and rpath is set.