Use GitHub Actions for Windows CI
Travis doesn't currently support python on Windows

This workflow was based on:
- https://hynek.me/articles/python-github-actions/
- https://github.com/actions/starter-workflows/pull/197

Currently this only _installs_ the extension and does not exercise any of
the tests. As basic installation check is performed by printing the
module's version. Some of the platform differences between windows and
*nix cause failures that originate in the test code and not the
extension. These will be fixed in a later commit.

Additionally, we configure the `checkout` action to clone the full
history in order for `setuptools_scm` to work correctly.

The built distribution (i.e., `bdist`) is persisted for use in manual
functional testing.